requirements.txt
annotated-types==0.6.0
anyio==3.7.1
azure-core==1.29.5
azure-identity==1.15.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
distlib==0.3.8
dnspython==2.6.1
fastapi==0.104.1
filelock==3.13.1
h11==0.14.0
hiredis==2.3.2
idna==3.4
loguru==0.7.2
more-itertools==9.1.0
msal==1.25.0
msal-extensions==1.0.0
pipenv==2023.11.15
platformdirs==4.1.0
portalocker==2.8.2
psycopg2==2.9.9
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.0.3
pydantic_core==2.3.0
PyJWT==2.8.0
pymongo==4.6.2
pyodbc==5.0.1
python-dotenv==1.0.0
python-ulid==1.1.0
pywin32==306
redis==4.6.0
redis-om==0.2.1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
types-pyOpenSSL==24.0.0.20240130
types-redis==4.6.0.20240218
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
virtualenv==20.25.0
win32-setctime==1.1.0
